Fix the logic for allChecksSucceeded
The logic originally only worked if there was only a single file for a
consignment. If there's more than one and any one has no status errors,
then it was returning allChecksSucceeded true.

I've changed this so it checks to make sure there are statuses and that
none of them are not Success. I've added a couple of tests to check for
this as well.