TDR-240 Return the same error for non existent consignment.
When we validate that the user owns the consignment, if the consignment
doesn't exist, we return "Invalid consignment Id" and if they don't have
permissions, then we get user xxx does not have accesss...
It's good practice to return the same authorisation error message
whether a resource exists or not to prevent people finding out which
resources exist by checking for the errors. I've changed it so you get
the same message and updated the tests which were checking for it.