Add functionality to update consignment status upon completion of upload.
These changes add repository and service methods alongside fields to create an endpoint to enable us to update the consignment status when a users' upload has completed.
I have created the appropriate tests alongside these in order to ensure functionality.

The consignment status endpoint is named `updateConsignmentStatusUploadComplete` to be very specific rather than a general update consignment status endpoint. This is to ensure that the aim of the endpoint is clear to those using it, in addition to keeping things very specific whilst we only have one use case for updating the consignment status. Once we have more use cases for updating consignment status, it may be more appropriate to make the endpoint more general.