Add graphql schema and task to check it matches the generated one.
I've created a custom sbt task which generates the graphql schema and
then compares it to the schema.graphql in the root of the project. If
they don't match, then the build fails.

This means that if you make changes to the API which change the schema,
you will need to keep this file up to date in order for the test build
to pass.