Add ConsignmentStatus service method and ConsignmentStatusFields
These changes allow us to utilise the consignmentStatusService method by calling the updateConsignmentStatusUploadComplete mutation field.
This has been done very specific to it's purpose rather than adding a more general 'updateConsignmentStatus' endpoint in order to make its purpose clearer when used within the FE.
The fields file has also been added to the GraphQlTypes in order for it to be used properly.