Catch all exceptions and log them.
I think what is happening here is that the exceptions are getting thrown
up the stack but aren't actually being caught anywhere so in the end,
the stack trace is just being printed to stderr.

I've changed the exception handler to catch anything not already
specified, to log the error and then to return an UNKNOWN_ERROR code.
I've done this so we're not returning specifics of every error to the
client but we could change this to the specific exception message if we
wanted to.