Move DbConnection.db from the Routes constructor.
What I think is happening is that DbConnection.db is being called when
the API starts which is caching the current IAM password but then it's
never being called again so we never get an updated passsword. I've
removed it from the Routes constructor and added called it within the
route definition.

I've moved the test which runs against the closed database out of the
RouteAuthenticationSpec as it's not possible to do it in there now and
added it into the FullHealthCheckServiceSpec class.