Use base alpine image with jdk 15 installed.
We have tried upgrading busybox on the openjdk image but it seems to be
on an older version of alpine and there are still vulnerabilities in the
latest version of busybox for that version.

Instead of using the openjdk image, I've changed this to use the
standard alpine image and installed openjdk 15 from the alpine edge
repository.

I've tested this on integration and it's working.