Start the process of making it difficult to add a consignment status that isn't 'InProgress'
From what I've concluded, when you add a consignment status value, it's always going to be 'InProgress', i.e., you are never going to add a
status value that is 'Completed'; to set it to 'Completed', you'd use the 'updateConsignmentStatus' method, so the default value has now been
set to 'InProgress' (statusValue: String='InProgress'). Once the 'addTransferAgreement' service method has been removed, then it can be set
as a fixed value in the 'ConsignmentRow'.