Add directory metadata.
When we add the files and directories, we were putting entries into the
"File" table for folders and directories but we were only putting
metadata in for files.

We need to include folders in the metadata csv in the Bagit package so
the easiest way is to add entries into the "FileMetadata" table. I've
added in the static metadata and included the full file path in the
ClientSideOriginalFilePath entry which should make life a bit easier for
the custom metadata as well.