Add a fork test job
PRs from forks like those that scala steward creates don't have acess to
secrets because they could change the action yaml files to print them
out.

This job uses pull_request_target which gets the yaml for the actions
from the master branch rather than the fork branch.

This won't work properly for our normal PRs because we won't be able to
make changes to the yaml files ourselves so we need a separate job that
only runs on tyeh update/ branches which scala steward creates.