Add mergify yaml and scala steward config.
I've signed this repo up to get the scala steward updates. We also want
these to be auto merged because we don't want more dependency updates to
worry about.

I've added in the mergify config and got scalastyle to ignore the
consignment-api-db changes as there's an ancient version in maven.

I've also change the test trigger from pull_request to
pull_request_target. This means that the code for the github actions
always comes from master but this is fine because there's a push trigger
as well which will catch any errors if we do make changes to the action
files.