Fix mismatched file ids
Originally, we were creating files and then assigning an arbitrary file
ID to add the metadata. This worked because there was nothing in the
file table which identified the file it came from.

We now have a file name but we're still arbitrarily assigning IDs to the
FileMetadata table so sometimes, the file name and the original path
don't match up.

I've updated this to create the file rows and the fileMetadata rows
together with each loop through the values from the generateNodes
method.

I've added a test to check that the file name and original path match
and updated a couple of other tests where this has broken them.