Add a total items field.
In this query, it is populated with total files which is used on the
front end to display the total number of files for that folder.

I kept the name slightly more generic because this case class can be
used for other non-file things.

Instead of writing whole new tests I've added a couple of assertions
about the new totalItems field.