Return an error for non judgment files
The allChecksSucceeded method was only returning an error for the FFID
status if it was contained in DisallowedPuids but that table doesn't
have NonJudgmentFormat in the table so if the status was
NonJudgmentFormat, it was returning allChecksSucceeded=true.

This checks for the statuses in that table and the NonJudgmentFormat.
I've added in a unit test to check for it.