Add redacted files query
There is only a single query here with an option to only filter by the
rows where the original file ID is null.

If there is a row where it is null, this means that there is a redacted
file with no corresponding original file. For the allChecksSucceeded
call, we check that there are no rows like this.

The same query is also used in the query used by the consignment export
but in this case, we send the file ID so this can be added to the
metadata export.

I've added in tests. For the file repository, I've tried to think of
every combination that would come up but I've probably missed some so
suggestions welcome.