Changes to only run this when requested
I've used a Sangria Projector which provides the arguments that have
been passed in the graphql query.

With this, we can check if originalFilePath has been passed. If it has,
then we get the file metadata as usual and update it with any original
file paths from the redacted files.

If it has not been requested, it gets the original file metadata but
without checking for redacted files.

The repository method now takes a list of file IDs and filters them if
the list is populated. This means that if any filters are applied on the
list of files, it will only search for those files.