Make ConsignmentStatus' statusValue optional but only for 'Upload'
Instead of creating a new endpoint I adapted the updateConsignmentStatus to take 'statusValue' as an optional parameter.
Although it's optional, it will only be optional for 'Upload' for now (there is logic to check this).
This way, when the files are uploaded, the frontend can call updateConsignmentStatus and then whether the upload is 'Complete'
or 'CompletedWithIssues' can be determined by the API (via the fileStatuses) rather than the frontend