Create a new multiple file metadata endpoint.
The existing updateBulkFileMetadata adds the same metadata values to
multiple files. For the new backend checks, we need a way of adding
different property values to multiple files.

I've added this in and updated the tests to point to the new one.
It's only the api update lambda that uses this endpoint so once this is
all deployed, we can remove the old one.