Add multiple file statuses
I've added in a new endpoint to allow us to add multiple file statuses
as part of the new backend checks.

I've deprecated the old method and updated the tests to use the new one

I've not changed the existing mutation types because the last time I did
that it broke things.

I've removed the restrictions on the status type only allowing Upload
statuses because we don't want that now.