Use file statuses for checks count.
At the moment, we're checking the relevant metadata fields to see if the
checks have completed. The problem is that if they fail, there is no
metadata so the checks page just hangs.

This checks the file statuses instead of the metadat itself because
these will be set to failed if there are any failures.