New metadata input validation tag
Directories cannot have additional metadata assigned to them.

New validator defensively checks for any directory ids in the input and throws an error.

Doing the check with a tag:
a) prevents any directories being updated with additional metadata and provides an error
b) reuses an existing DB call for the checking authorisation reducing the amount of DB calls needs in the overall updating process.

Authorisation exception is given precendance over data input exceptions as the input cannot be used if the user is not authorised.

The validator check for directories can be re-visited if:
a) its decided directories can have additional metadata assigned to them
b) need to pass in directories ids for bulk metadata editing